bugfixing: merged words after html cleaning and NAs for cleaned non-files
fixed a bug that caused seperate words to join to a single one in html cleaning. All characters bewteen the words may have been
removed. Now, tags are replaced by spaces, and multiple spaces are replaced by a single one in the end.

When the cleaned file cannot be written, <NA> is now stated as entry rather than NaN.